requirements.txt
anthropic
openai
python-dotenv
colorama
pygments
tavily-python
setuptools